+++ spec/jobs/dummy_job.rb
@@ -1,7 +1,8 @@
 module Jobs
   class DummyJob < Jobi::Job
     options queue_name: :dummies,
-            consumers: 2
+            consumers: 2,
+            persist: true
 
     def initialize(first:, second:)
       @first = first